backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
